+++ test/packaging/run.sh
@@ -15,6 +15,7 @@ for v in "${here}/verify_layout.sh" \
          "${here}/verify_ready_check.sh" \
          "${here}/verify_systemd_units.sh" \
          "${here}/verify_lifecycle_scripts.sh" \
+         "${here}/verify_macos_homebrew_lifecycle.sh" \
          "${here}/verify_descriptor.sh" \
          "${here}/verify_installer.sh"; do
   ran=$((ran + 1))